type fix

+++ src/test_skpk/main.cpp
@@ -125,7 +125,7 @@ std::string EncodeBase58(const unsigned char* pbegin, const size_t sz)
         zeroes++;
 
     // Allocate enough space in big-endian base58 representation.
-    int size = sz * 138 / 100 + 1; // log(256) / log(58), rounded up.
+    size_t size = sz * 138 / 100 + 1; // log(256) / log(58), rounded up.
     std::vector<unsigned char> b58(size);
     // Process the bytes.
     for ( ; pbegin != pend; pbegin++)
@@ -186,7 +186,7 @@ bool DecodeBase58(const char* psz, std::vector<unsigned char>& vch)
         zeroes++;
 
     // Allocate enough space in big-endian base256 representation.
-    int size = strlen(psz) * 733 /1000 + 1; // log(58) / log(256), rounded up.
+    size_t size = strlen(psz) * 733 /1000 + 1; // log(58) / log(256), rounded up.
     std::vector<unsigned char> b256(size);
     // Process the characters.
     static_assert(sizeof(mapBase58)/sizeof(mapBase58[0]) == 256, "mapBase58.size() should be 256"); // guarantee not out of range
